refactor: 친구 관계, 친구 초대 도메인 패키지 분리 #37
