+++ src/main/java/site/praytogether/pray_together/domain/friend/domain/friend_invitation/FriendInvitationRepository.java
@@ -1,4 +1,4 @@
-package site.praytogether.pray_together.domain.friend.domain;
+package site.praytogether.pray_together.domain.friend.domain.friend_invitation;
 
 public interface FriendInvitationRepository {
   FriendInvitation save(FriendInvitation friendInvitation);
