+++ src/main/java/site/praytogether/pray_together/domain/friend/domain/friendship/FriendshipService.java
@@ -1,4 +1,4 @@
-package site.praytogether.pray_together.domain.friend.domain;
+package site.praytogether.pray_together.domain.friend.domain.friendship;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
