+++ src/main/java/site/praytogether/pray_together/domain/friend/infratstucture/persistence/FriendInvitationJpaRepository.java
@@ -1,7 +1,7 @@
 package site.praytogether.pray_together.domain.friend.infratstucture.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.praytogether.pray_together.domain.friend.domain.FriendInvitation;
+import site.praytogether.pray_together.domain.friend.domain.friend_invitation.FriendInvitation;
 
 public interface FriendInvitationJpaRepository extends JpaRepository<FriendInvitation, Long> {
