+++ src/main/java/site/praytogether/pray_together/domain/friend/infratstucture/persistence/FriendInvitationRepositoryImpl.java
@@ -2,8 +2,8 @@ package site.praytogether.pray_together.domain.friend.infratstucture.persistence
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import site.praytogether.pray_together.domain.friend.domain.FriendInvitation;
-import site.praytogether.pray_together.domain.friend.domain.FriendInvitationRepository;
+import site.praytogether.pray_together.domain.friend.domain.friend_invitation.FriendInvitation;
+import site.praytogether.pray_together.domain.friend.domain.friend_invitation.FriendInvitationRepository;
 
 @Repository
 @RequiredArgsConstructor
