+++ src/main/java/site/praytogether/pray_together/domain/friend/infratstucture/persistence/FriendshipRepositoryImpl.java
@@ -2,7 +2,7 @@ package site.praytogether.pray_together.domain.friend.infratstucture.persistence
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import site.praytogether.pray_together.domain.friend.domain.FriendshipRepository;
+import site.praytogether.pray_together.domain.friend.domain.friendship.FriendshipRepository;
 
 @Repository
 @RequiredArgsConstructor
